Rename hessian_optimization -> cached_hessian
