+++ tlm_adjoint/__init__.py
@@ -18,22 +18,23 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
 
-from .caches import *                # noqa: F401
-from .eigendecomposition import *    # noqa: F401
-from .equation import *              # noqa: F401
-from .equations import *             # noqa: F401
-from .fixed_point import *           # noqa: F401
-from .functional import *            # noqa: F401
-from .hessian import *               # noqa: F401
+from .cached_hessian import *      # noqa: F401
+from .caches import *              # noqa: F401
+from .eigendecomposition import *  # noqa: F401
+from .equation import *            # noqa: F401
+from .equations import *           # noqa: F401
+from .fixed_point import *         # noqa: F401
+from .functional import *          # noqa: F401
+from .hessian import *             # noqa: F401
+from .interface import *           # noqa: F401
+from .linear_equation import *     # noqa: F401
+from .manager import *             # noqa: F401
+from .optimization import *        # noqa: F401
+from .overloaded_float import *    # noqa: F401
+from .storage import *             # noqa: F401
+from .tangent_linear import *      # noqa: F401
+from .tlm_adjoint import *         # noqa: F401
+from .verification import *        # noqa: F401
+
 from .hessian import GeneralGaussNewton as GaussNewton  # noqa: F401
-from .hessian import GeneralHessian as Hessian  # noqa: F401
-from .hessian_optimization import *  # noqa: F401
-from .interface import *             # noqa: F401
-from .linear_equation import *       # noqa: F401
-from .manager import *               # noqa: F401
-from .optimization import *          # noqa: F401
-from .overloaded_float import *      # noqa: F401
-from .storage import *               # noqa: F401
-from .tangent_linear import *        # noqa: F401
-from .tlm_adjoint import *           # noqa: F401
-from .verification import *          # noqa: F401
+from .hessian import GeneralHessian as Hessian          # noqa: F401
